ZIndex: Report computed z-index as specified, ignoring stacking context.
This patch changes the behavior of computed style z-index to
comply with spec: it reports the value as specified instead of
changing it to auto if the style does not induce a stacking context.

Note that since z-index only applies on positioned elements, and
grid and flex children, this patch introduces an EffectiveZIndex() which
is set to 0 if the z-index does not apply. This is so that code relying
on z-index will behave as it did before. However, ZIndex() now reflects
the value that was specified.

R=futhark@chromium.org, pdr@chromium.org
Fixed: 1023232
Change-Id: Ib11467a9818b746c72bebbb167223f1bb41a5f9f